api/requirements.txt

Flask==2.3.3
Flask-Cors==4.0.0
pandas==2.1.0
numpy==1.26.0
gunicorn==21.2.0
openai==0.28.0
python-dotenv==0.21.1
